+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add suffix to download file for csv typed response #502
 * ResourceLoader - support any object type #497
 * Further simplify password field processing #491
+* Encrypt sensitive data in persistent storage #490
 * Do not `Set-Cookie` for session and flash if there is no state #484
 * Make `act.metric.Timer` be a `Closable` #483
 * Simplify measuring of method execution #482